split NameAndFundingSourceForm into two components
