+++ web/src/pages/apd/activities/overview/NameAndFundingSource/index.js
@@ -1,5 +1,3 @@
-import NameAndFundingSourceForm from './NameAndFundingSourceForm';
 import NameAndFundingSourceReview from './NameAndFundingSourceReview';
-import NameForm from './NameForm';
 
-export { NameAndFundingSourceForm, NameAndFundingSourceReview, NameForm };
+export { NameAndFundingSourceReview };
