+++ web/src/pages/apd/activities/overview/NameForm.test.js
@@ -6,31 +6,27 @@ import {
   waitFor
 } from 'apd-testing-library';
 
-import { plain as NameAndFundingSourceForm } from './NameAndFundingSourceForm';
+import { plain as NameForm } from './NameForm';
 
 const defaultProps = {
   index: 1,
   item: {
-    fundingSource: 'Uncle Scrooge',
     key: 'key 1',
     name: 'Buying bikes for Huey, Dewey, and Louie'
   },
-  setFundingSource: jest.fn(),
   setName: jest.fn()
 };
 
 const setup = async (props = {}) => {
   // eslint-disable-next-line testing-library/no-unnecessary-act
   const utils = await act(async () =>
-    renderWithConnection(
-      <NameAndFundingSourceForm {...defaultProps} {...props} />
-    )
+    renderWithConnection(<NameForm {...defaultProps} {...props} />)
   );
   await waitFor(() => screen.findByText(/Activity Name/i));
   return utils;
 };
 
-describe('the NameAndFundingSourceForm component', () => {
+describe('the NameForm component', () => {
   beforeEach(() => {
     jest.resetAllMocks();
   });
